cloud-run-streaming/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.52.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.17.0
python-docx==0.8.11
pypdf==3.17.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0 